+++ src/suou/quart.py
@@ -14,4 +14,70 @@ This software is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 """
 
-# TODO everything
+from __future__ import annotations
+
+from flask import current_app
+from quart import Quart, request, g
+from quart_schema import QuartSchema
+
+from suou.http import WantsContentType
+
+from .i18n import I18n
+from .itertools import makelist
+
+def add_i18n(app: Quart, i18n: I18n, var_name: str = 'T', *,
+        query_arg: str = 'lang', default_lang = 'en'):
+    '''
+    Integrate a I18n() object with a Quart application:
+    - set g.lang
+    - add T() to Jinja templates
+
+    XXX UNTESTED
+    '''
+    def _get_lang():
+        lang = request.args.get(query_arg)
+        if not lang:
+            for lp in request.headers.get('accept-language', 'en').split(','):
+                l = lp.split(';')[0]
+                lang = l
+                break
+            else:
+                lang = default_lang
+        return lang
+    
+    @app.context_processor
+    def _add_i18n():
+        return {var_name: i18n.lang(_get_lang()).t}
+
+    @app.before_request
+    def _add_language_code():
+        g.lang = _get_lang()
+
+    return app
+
+
+def negotiate() -> WantsContentType:
+    """
+    Return an appropriate MIME type for content negotiation.
+    """
+    if 'application/json' in request.accept_mimetypes or any(request.path.startswith(f'/{p.strip('/')}/') for p in current_app.config.get('REST_PATHS')):
+        return WantsContentType.JSON
+    elif request.user_agent.string.startswith('Mozilla/'):
+        return WantsContentType.HTML
+    else:
+        return WantsContentType.PLAIN
+
+
+def add_rest(app: Quart, *bases: str, **kwargs) -> QuartSchema:
+    """
+    Construct a REST ...
+
+    The rest of ...
+    """
+
+    schema = QuartSchema(app, **kwargs)
+    app.config['REST_PATHS'] = makelist(bases, wrap=False)
+    return schema
+
+
+__all__ = ('add_i18n', 'negotiate', 'add_rest')